revert get support languages change
